pyproject.toml
[project]
name = "picbot"
version = "0.1.0"
description = ""
authors = [
    {name = "BoBkiNN",email = "78136575+BoBkiNN@users.noreply.github.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "disnake (>=2.11.0,<3.0.0)",
    "serpapi (>=0.1.5,<0.2.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
